requirements.txt
python-engineio==3.13.2
python-socketio==4.6.0
flask-socketio=4.3.1
celery==4.4.2
pytest
coverage
